refactor(server): extract fleet orchestration routes to routers/orchestration.py (issue #359)
- Create backend/orchestration_audit.py with shared lock, NDJSON audit log
  helpers (load_orchestration_audit, append_orchestration_audit, migration),
  and a corruption counter exposed via get_audit_log_corrupt_total()
- Create backend/routers/orchestration.py with 12 extracted routes:
  GET/POST /api/audit, /api/fleet/audit, /api/fleet/orchestration,
  /api/fleet/orchestration/dispatch, /api/fleet/orchestration/deploy,
  /api/fleet/control/{action}, /api/fleet/schedule, /api/fleet/capacity,
  /api/fleet/nodes, /api/fleet/hardware, /api/fleet/nodes/{node_name}/system
- Wire router via app.include_router and set_dependencies() in server.py
- Remove old route handlers and audit functions from server.py
- Keep _fleet_control_local, _remote_fleet_control, _get_fleet_nodes_impl
  as server.py helpers injected via set_dependencies() (no logic change)
- Remove now-redundant _orchestration_audit_lock from server.py
  (lock lives exclusively in orchestration_audit.py, resolves #344 concern)
